requirements.txt
Flask_Login==0.5.0
alembic==1.4.2
Flask_Migrate==2.5.3
SQLAlchemy==1.3.18
Flask_WTF==0.14.3
Werkzeug==0.16.0
Flask==1.0.2
Flask_Dropzone==1.5.4
Flask_SQLAlchemy==2.4.4
WTForms==2.3.3
Flask_Uploads==0.2.1
Pillow==7.2.0
